removed some more status functions, updated script converter & syntax checker to be conform with the new syntax
git-svn-id: http://eathena.googlecode.com/svn/branches/eapp@9002 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/basics/baseparser.h
@@ -208,8 +208,8 @@ class CParseInput : public allocator_file<char>
 public:
 	///////////////////////////////////////////////////////////////////////////
 	// position marker
-	unsigned short	line;
-	unsigned short	column;
+	unsigned int	line;
+	unsigned int	column;
 
 	CParseInput() : line(1),column(0)	{}
 	virtual ~CParseInput()				{}
